+++ lib/requests_inspector.dart
@@ -7,6 +7,7 @@ export 'src/requests_inspector_widget.dart';
 export 'src/enums/requests_methods.dart';
 export 'src/enums/show_inspector_on_enum.dart';
 export './src/requests_inspector_interceptor.dart';
+export 'src/helpers/sensitive_data_masker.dart';
 export 'src/graphql_inspector_link.dart';
 export 'src/request_stopper_editor_dialog.dart';
 export 'src/response_stopper_editor_dialog.dart';